feat(functions): made functions public
